Move `createami` implementation to its own module
